requirements.txt
streamlit
pandas
lxml
Pillow
tesseract
pytesseract